Refactor DoMpcConfig for improved model handling and TVP updates
- Updated the `create_mpc_controller` method to accept the model as a parameter.
- Changed variable types for reference trajectory parameters and disturbances to use time-varying parameters (TVP).
- Enhanced the TVP function to correctly update reference trajectory parameters during simulation.
- Refactored numpy array handling for consistency in trajectory calculations in `trajectory_ref.py`.

+++ src/system/trajectory_ref.py
@@ -118,8 +118,8 @@ class Trajectory:
 
         # Calculate attitude
         phi_d = 0.0  # Maintain zero roll
-        theta_d = np.arctan2(-vel[2], np.sqrt(vel[0] ** 2 + vel[1] ** 2))  # Pitch angle
-        psi_d = np.arctan2(vel[1], vel[0])  # Yaw angle
+        theta_d = float(np.arctan2(-vel[2], np.sqrt(vel[0] ** 2 + vel[1] ** 2)) ) # Pitch angle
+        psi_d = float(np.arctan2(vel[1], vel[0]) ) # Yaw angle
         gamma_d = np.array([phi_d, theta_d, psi_d])
 
         # Use numerical differentiation to get attitude derivatives
